Fix bug in the enso processing module
The commit contains changes to fix the link directory name for accessing
the files needed for ENSO diagnostics. In addition, the variables
including sea surface height(ssh) and energy fluxes (thf) may not
readily available. A relax rule for the sucess checks was added so that
the diagnostics was not marked as failed.